Added ITEM_OUTLINE_BLACK_SECONDARY
This should help item recognition for items/slots that do not have the
primary black outline color. The only function changed is isItemIn(box).
This function was the only one changed because this functions seems to
be a function meant for global use. However, other functions, such as
banking functions that check for item outline are meant for specific
purposes where an item outline of the primary black color should render
accurate results. If other files were to be changed, then isItemIn(box)
can be used.