function TRSChatBox.findTextBoxes(search: string): TBoxArray;
    function TRSChatBox.findTextBoxes(search: string): TBoxArray;

Returns the TBoxArray of the type/color of text it **searches** for.
Usefull for accepting trades or adding certain players to ignore. They
are sorted from top to bottom (old to new). The possible types of text
to search for can be found at the top of the chatbox.simba file.

.. note::

    - by Thomas
    - Last Updated: 11 December by Thomas

Example:

.. code-block:: pascal

    trades := chatBox.findTextBoxes(CHATBOX_TRADES);
    if length(trades) then
      mouseBox(trades[high(trades)], MOUSE_RIGHT);