Update time.simba
Added an optional parameter 'Value' to waitFunc and waitTypeFunc, so that they will wait for the function to match the Value instead of just waiting for the function to be true. This allows things such as `waitTypeFunc(@bankScreen.isOpen, 50, 2000, false);` which waits for the bankScreen to be closed.

This does not break backwards compatibility; The extra parameter is optional, and defaults to true so the function will behave the same for existing implementations.